% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-builder.R
\name{cb_reset}
\alias{cb_reset}
\title{Reset a context builder}
\usage{
cb_reset(builder)
}
\arguments{
\item{builder}{A \code{securecontext_context_builder}.}
}
\value{
Reset builder.
}
\description{
Removes all added content.
}
