% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-builder.R
\name{context_reset}
\alias{context_reset}
\alias{cb_reset}
\title{Reset a context builder}
\usage{
context_reset(builder)

cb_reset(...)
}
\arguments{
\item{builder}{A \code{securecontext_context_builder}.}

\item{...}{Arguments passed to \code{\link[=context_reset]{context_reset()}}.}
}
\value{
Reset builder.
}
\description{
Removes all added content.
}
\examples{
cb <- context_builder(max_tokens = 100)
cb <- context_add(cb, "some text")
cb <- context_reset(cb)
length(cb@items)
}
