% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{decrypt_raw}
\alias{decrypt_raw}
\title{Decrypt raw data with AES-256-CBC}
\usage{
decrypt_raw(data, key)
}
\arguments{
\item{data}{Encrypted raw vector (IV prepended).}

\item{key}{Raw 32-byte key.}
}
\value{
Decrypted raw vector.
}
\description{
Decrypt raw data with AES-256-CBC
}
\keyword{internal}
