% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{encrypt_raw}
\alias{encrypt_raw}
\title{Encrypt raw data with AES-256-CBC}
\usage{
encrypt_raw(data, key)
}
\arguments{
\item{data}{Raw vector to encrypt.}

\item{key}{Raw 32-byte key.}
}
\value{
Encrypted raw vector (IV prepended).
}
\description{
Encrypt raw data with AES-256-CBC
}
\keyword{internal}
