% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{is_document}
\alias{is_document}
\title{Test if object is a document}
\usage{
is_document(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Test if object is a document
}
\examples{
doc <- document("Hello world")
is_document(doc)
is_document("not a doc")
}
