% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-builder.R
\name{securecontext_context_builder}
\alias{securecontext_context_builder}
\title{S7 class for securecontext context builders}
\usage{
securecontext_context_builder(max_tokens = integer(0), items = list())
}
\arguments{
\item{max_tokens}{Integer, maximum number of tokens.}

\item{items}{List of content items with priority.}
}
\description{
S7 class for securecontext context builders
}
\examples{
cb <- context_builder(max_tokens = 100)
cb@max_tokens
}
