% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{securecontext_embedder}
\alias{securecontext_embedder}
\title{S7 class for securecontext embedders}
\usage{
securecontext_embedder(embed_fn = function() NULL, dims = integer(0))
}
\arguments{
\item{embed_fn}{A function taking a character vector and returning a numeric
matrix.}

\item{dims}{Integer, the dimensionality of the embedding space.}
}
\value{
A \code{securecontext_embedder} S7 object.
}
\description{
S7 class for securecontext embedders
}
\examples{
emb <- embed_tfidf(c("hello world", "goodbye world"))
emb@dims
}
