% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retriever.R
\name{securecontext_retriever}
\alias{securecontext_retriever}
\title{S7 class for securecontext retrievers}
\usage{
securecontext_retriever(store = NULL, embedder = NULL)
}
\arguments{
\item{store}{A \link{vector_store} object.}

\item{embedder}{A \code{securecontext_embedder} object.}
}
\value{
A \code{securecontext_retriever} S7 object.
}
\description{
S7 class for securecontext retrievers
}
\examples{
emb <- embed_tfidf(c("cat sat on mat", "dog ran in park"))
vs <- vector_store$new(dims = emb@dims)
ret <- retriever(vs, emb)
ret@embedder
}
